Refactors cosmoz-page-router
